core, datasource, runtime, store: Fix build errors in tests
